[Refactor] Restructure body models and registrants (#56)
* SMPLify is moved from core/parametric_models to models/registrants to share the same module builder

* SMPL wrappers are moved to models/body_models to accomodate more elaborate inheritances
