--- mmhuman3d/core/parametric_model/__init__.py
@@ -1,4 +1,0 @@
-from .builder import build_registrant
-from .smplify import SMPLify, SMPLifyX
-
-__all__ = ['build_registrant', 'SMPLify', 'SMPLifyX']
