+++ mmhuman3d/models/__init__.py
@@ -1,5 +1,6 @@
 from .architectures import *  # noqa: F401,F403
 from .backbones import *  # noqa: F401,F403
+from .body_models import *  # noqa: F401,F403
 from .builder import (
     ARCHITECTURES,
     BACKBONES,
@@ -20,6 +21,7 @@ from .discriminators import *  # noqa: F401,F403
 from .heads import *  # noqa: F401,F403
 from .losses import *  # noqa: F401,F403
 from .necks import *  # noqa: F401,F403
+from .registrants import *  # noqa: F401,F403
 
 __all__ = [
     'BACKBONES', 'LOSSES', 'ARCHITECTURES', 'HEADS', 'BODY_MODELS', 'NECKS',
