+++ mmhuman3d/models/body_models/__init__.py
@@ -1,0 +1,6 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+
+from .smpl import SMPL, GenderedSMPL, HybrIKSMPL
+from .smplx import SMPLX
+
+__all__ = ['SMPL', 'GenderedSMPL', 'HybrIKSMPL', 'SMPLX']
