+++ mmhuman3d/models/registrants/__init__.py
@@ -1,0 +1,4 @@
+from .smplify import SMPLify
+from .smplifyx import SMPLifyX
+
+__all__ = ['SMPLify', 'SMPLifyX']
