+++ mmhuman3d/models/utils/__init__.py
@@ -1,4 +1,8 @@
+from .fits_dict import FitsDict
+from .inverse_kinematics import batch_inverse_kinematics_transform
 from .res_layer import ResLayer, SimplifiedBasicBlock
-from .smpl import SMPL, HybrIKSMPL
 
-__all__ = ['ResLayer', 'SimplifiedBasicBlock', 'SMPL', 'HybrIKSMPL']
+__all__ = [
+    'FitsDict', 'ResLayer', 'SimplifiedBasicBlock',
+    'batch_inverse_kinematics_transform'
+]
