+++ tests/test_models/test_heads/test_hybrik_forward.py
@@ -7,7 +7,7 @@ import pytest
 import torch
 
 from mmhuman3d.models import HybrIK_trainer, HybrIKHead
-from mmhuman3d.models.utils import HybrIKSMPL
+from mmhuman3d.models.builder import build_body_model
 from mmhuman3d.models.utils.inverse_kinematics import (
     batch_get_3children_orient_svd,
     batch_get_pelvis_orient,
@@ -42,9 +42,12 @@ def test_HybrIK_head():
     # initialize models
     head = HybrIKHead(
         smpl_mean_params=osp.join(tmpdir.name, 'h36m_mean_beta.npy'))
-    smpl = HybrIKSMPL(
-        model_path='data/body_models/smpl',
-        extra_joints_regressor=osp.join(tmpdir.name, 'J_regressor_h36m.npy'))
+    smpl = build_body_model(
+        dict(
+            type='HybrIKSMPL',
+            model_path='data/body_models/smpl',
+            extra_joints_regressor=osp.join(tmpdir.name,
+                                            'J_regressor_h36m.npy')))
 
     if torch.cuda.is_available():
         head = head.cuda()
