+++ tests/test_smplify.py
@@ -2,8 +2,7 @@ import mmcv
 import numpy as np
 import torch
 
-from mmhuman3d.core.parametric_model.builder import build_registrant
-from mmhuman3d.models.builder import build_body_model
+from mmhuman3d.models.builder import build_body_model, build_registrant
 
 body_model_load_dir = 'data/body_models'
 batch_size = 2
