+++ tools/smplify.py
@@ -6,8 +6,8 @@ import numpy as np
 import torch
 
 from mmhuman3d.core.conventions.keypoints_mapping import convert_kps
-from mmhuman3d.core.parametric_model import build_registrant
 from mmhuman3d.core.visualization.visualize_smpl import visualize_smpl_pose
+from mmhuman3d.models.builder import build_registrant
 
 
 def parse_args():
